Migrate base modification code out of sam.c
Sam.c is already a behemoth over 6,000 lines long.  It's hard to find
things in it, and splitting by domain (with a sam prefix) makes life
easier as a developer.

Note: when sorted by file size, the split off code is already close to
the median, so this isn't creating pointlessly small files.  The
functions are still declared in sam.h there is no API change.

This commit changes no functionality as it's simply code migration.
